wip: refact tests
